Refactor ipc handles in index
